requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==
chardet==4.0.0
click==8.0.3
cloud-init==
colorama==0.4.4
command-not-found==0.3
configobj==
constantly==15.1.0
contourpy==1.3.2
cryptography==
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
matplotlib==3.10.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==2.1.2
nvidia-ml-py==13.580.82
oauthlib==
packaging==25.0
pandas==2.3.3
pillow==
psutil==7.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynvml==13.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sympy==1.13.3
systemd-python==234
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2025.2
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
